Updated bunch of commands
Broke commands into more categories, quick dirty fix for help menu

+++ commands/general/manga.js
@@ -33,17 +33,13 @@ class Manga extends Command {
     try {
       const { data } = await kitsu.fetch("manga", { filter: { text: args.join("-") } });
       msg = await msg.edit(`Okay I found 5 possible matches which do you want to see? (just write the first number, it will be canceled after 60 seconds)${this.makeTitles(data)}`);
-      const collected = await message.channel.awaitMessages(filter, {
-        max: 20,
-        maxMatches: 1,
-        time: 60000,
-        errors: ["time"]
-      });
+      const collected = await message.channel.awaitMessages(filter, { max: 20, maxProcessed: 1, time: 60000, errors: ["time"] });
       const returnMessage = collected.first();
-      await returnMessage.delete();
       const index = Number(returnMessage.content) - 1;
+      if (message.channel.permissionsFor(this.client.user).has("MANAGE_MESSAGES")) await returnMessage.delete();
       await msg.edit(`**Title JP:** ${data[index].titles.en_jp}\n**Title English:** ${data[index].titles.en}\n**Type:** ${data[index].subtype}\n**Start Date:** ${data[index].startDate}\n**End Date:** ${data[index].endDate || "in Progress"}\n**PopularityRank:** ${data[index].popularityRank}\n**Link:** <https://kitsu.io/manga/${data[index].id}>\n**Synopsis:** ${data[index].synopsis}`);
     } catch (error) {
+      console.log(error);
       if (error instanceof Collection) return message.reply("command canceled due timer");
       await msg.edit("I had a error while trying to fetch the data from Kitsu Sorry! did you spell the Manga name right?");
       await message.react("❓");
